Add clarification of how to set force_hostname
The original instructions told the user to log into Discourse Admin settings in order to set the force_hostname parameter so that the email validation URLs would be correct.  However, it is not possible to log in until email is validated.   This commit adds instructions for setting force_hostname through the console.