Load plugin settings.
Along the lines for loading the locale files for a plugin we should also load the settings.yml for a plugin.  Updated version 04f3e09 that moves the loading to a function. I do not understand why it was necessary to define the function physically earlier in the file as I thought ruby did two pass but that may only apply to functions and not stray code in the class.