Prevent rendering of excessive <li> wrapper around a category breadcrumb which already is a <li>.
The excessive <li> wrapper coming from the template is automatically closed by the browser during the HTML parsing step, causing an empty <li> being visible in the rendered source (preceding the first breadcrumb). This causes a visual bug when not using the default Discourse stylesheet, because the extra LI element pushes the content to the right and sometimes also renders an extra black pixel.