Traverse symlinks to plugins in dev mode when compiling stylesheets
When developing plugins it's useful to symlink the to the plugin directory from the discourse directory, since that means the two are separate git repos.

However, Dir.glob doesn't by default traverse symlinks. This change means that the SASS compilation caching detects when any of a plugin's files have changed.