Fix bug where create topic is suggested when insufficient permissions
When a user is viewing a category they do not have permission to create
a topic in, they are still prompted to create a topic if they are at the
bottom of the topics. This fixes the issue, as well as disabling the
New Topic button if they don't have permission to create a topic for
that category.